feat(klaviyo): inline email events into connection details
Drop the standalone KlaviyoEmailEventsModal and render the per-store
event matrix directly inside the connection details view. Cuts an extra
modal hop, makes per-store toggles immediately visible alongside the
connection metadata, and removes a separate component to keep in sync.
